requirements.txt
imageio==2.1.2
tqdm==4.26.0
torchsummary==1.5.1
tensorboardX==1.4
pillow==5.3.0
numpy==1.14.6
